+++ prov/shm/Makefile.include
@@ -3,6 +3,11 @@ _shm_files = \
 	prov/shm/src/smr_attr.c		\
 	prov/shm/src/smr_cq.c		\
 	prov/shm/src/smr_domain.c	\
+	prov/shm/src/smr_progress.c	\
+	prov/shm/src/smr_comp.c		\
+	prov/shm/src/smr_msg.c		\
+	prov/shm/src/smr_rma.c		\
+	prov/shm/src/smr_atomic.c	\
 	prov/shm/src/smr_ep.c		\
 	prov/shm/src/smr_fabric.c	\
 	prov/shm/src/smr_init.c		\
