+++ prov/shm/src/smr_ep.c
@@ -37,6 +37,10 @@
 #include "fi_iov.h"
 #include "smr.h"
 
+extern struct fi_ops_msg smr_msg_ops;
+extern struct fi_ops_tagged smr_tagged_ops;
+extern struct fi_ops_rma smr_rma_ops;
+extern struct fi_ops_atomic smr_atomic_ops;
 
 int smr_setname(fid_t fid, void *addr, size_t addrlen)
 {
@@ -105,113 +109,7 @@ static struct fi_ops_ep smr_ep_ops = {
 	.tx_size_left = fi_no_tx_size_left,
 };
 
-static int smr_tx_comp(struct smr_ep *ep, void *context, uint64_t flags,
-		       uint64_t err)
-{
-	struct fi_cq_tagged_entry *comp;
-	struct util_cq_err_entry *entry;
-
-	comp = ofi_cirque_tail(ep->util_ep.tx_cq->cirq);
-	if (err) {
-		if (!(entry = calloc(1, sizeof(*entry))))
-			return -FI_ENOMEM;
-		entry->err_entry.op_context = context;
-		entry->err_entry.flags = flags;
-		entry->err_entry.err = err;
-		entry->err_entry.prov_errno = -err;
-		slist_insert_tail(&entry->list_entry,
-				  &ep->util_ep.tx_cq->err_list);
-		comp->flags = UTIL_FLAG_ERROR;
-	} else {
-		comp->op_context = context;
-		comp->flags = flags;
-		comp->len = 0;
-		comp->buf = NULL;
-		comp->data = 0;
-	}
-	ofi_cirque_commit(ep->util_ep.tx_cq->cirq);
-	return 0;
-}
-
-static int smr_tx_comp_signal(struct smr_ep *ep, void *context, uint64_t flags,
-			      uint64_t err)
-{
-	int ret;
-
-	ret = smr_tx_comp(ep, context, flags, err);
-	if (ret)
-		return ret;
-	ep->util_ep.tx_cq->wait->signal(ep->util_ep.tx_cq->wait);
-	return 0;
-}
-
-static int smr_rx_comp(struct smr_ep *ep, void *context, uint64_t flags,
-		       size_t len, void *buf, void *addr, uint64_t tag,
-		       uint64_t data, uint64_t err)
-{
-	struct fi_cq_tagged_entry *comp;
-	struct util_cq_err_entry *entry;
-
-	comp = ofi_cirque_tail(ep->util_ep.rx_cq->cirq);
-	if (err) {
-		if (!(entry = calloc(1, sizeof(*entry))))
-			return -FI_ENOMEM;
-		entry->err_entry.op_context = context;
-		entry->err_entry.flags = flags;
-		entry->err_entry.tag = tag;
-		entry->err_entry.err = err;
-		entry->err_entry.prov_errno = -err;
-		slist_insert_tail(&entry->list_entry,
-				  &ep->util_ep.rx_cq->err_list);
-		comp->flags = UTIL_FLAG_ERROR;
-	} else {
-		comp->op_context = context;
-		comp->flags = flags;
-		comp->len = len;
-		comp->buf = buf;
-		comp->data = data;
-		comp->tag = tag;
-	}
-	ofi_cirque_commit(ep->util_ep.rx_cq->cirq);
-	return 0;
-}
-
-static int smr_rx_src_comp(struct smr_ep *ep, void *context, uint64_t flags,
-			   size_t len, void *buf, void *addr, uint64_t tag,
-			   uint64_t data, uint64_t err)
-{
-	ep->util_ep.rx_cq->src[ofi_cirque_windex(ep->util_ep.rx_cq->cirq)] =
-		(uint32_t) (uintptr_t) addr;
-	return smr_rx_comp(ep, context, flags, len, buf, addr, tag, data, err);
-}
-
-static int smr_rx_comp_signal(struct smr_ep *ep, void *context, uint64_t flags,
-			      size_t len, void *buf, void *addr, uint64_t tag,
-			      uint64_t data, uint64_t err)
-{
-	int ret;
-
-	ret = smr_rx_comp(ep, context, flags, len, buf, addr, tag, data, err);
-	if (ret)
-		return ret;
-	ep->util_ep.rx_cq->wait->signal(ep->util_ep.rx_cq->wait);
-	return 0;
-}
-
-static int smr_rx_src_comp_signal(struct smr_ep *ep, void *context, uint64_t flags,
-				  size_t len, void *buf, void *addr, uint64_t tag,
-				  uint64_t data, uint64_t err)
-{
-	int ret;
-
-	ret = smr_rx_src_comp(ep, context, flags, len, buf, addr, tag, data, err);
-	if (ret)
-		return ret;
-	ep->util_ep.rx_cq->wait->signal(ep->util_ep.rx_cq->wait);
-	return 0;
-
-}
-static int smr_verify_peer(struct smr_ep *ep, int peer_id)
+int smr_verify_peer(struct smr_ep *ep, int peer_id)
 {
 	int ret;
 
@@ -221,528 +119,6 @@ static int smr_verify_peer(struct smr_ep *ep, int peer_id)
 	ret = smr_map_to_region(&smr_prov, &ep->region->map->peers[peer_id]);
 
 	return (ret == -ENOENT) ? -FI_EAGAIN : ret;
-}
-
-static const uint64_t smr_tx_flags[] = {
-	[ofi_op_msg] = FI_SEND,
-	[ofi_op_tagged] = FI_SEND | FI_TAGGED,
-	[ofi_op_read_req] = FI_RMA | FI_READ,
-	[ofi_op_write] = FI_RMA | FI_WRITE,
-	[ofi_op_atomic] = FI_ATOMIC | FI_WRITE,
-	[ofi_op_atomic_fetch] = FI_ATOMIC | FI_WRITE | FI_READ,
-	[ofi_op_atomic_compare] = FI_ATOMIC | FI_WRITE | FI_READ,
-};
-
-static uint64_t smr_tx_comp_flags(uint32_t op)
-{
-	return smr_tx_flags[op];
-}
-
-static const uint64_t smr_rx_flags[] = {
-	[ofi_op_msg] = FI_RECV,
-	[ofi_op_tagged] = FI_RECV | FI_TAGGED,
-	[ofi_op_read_req] = FI_RMA | FI_REMOTE_READ,
-	[ofi_op_write] = FI_RMA | FI_REMOTE_WRITE,
-	[ofi_op_atomic] = FI_ATOMIC | FI_REMOTE_WRITE,
-	[ofi_op_atomic_fetch] = FI_ATOMIC | FI_REMOTE_WRITE | FI_REMOTE_READ,
-	[ofi_op_atomic_compare] = FI_ATOMIC | FI_REMOTE_WRITE | FI_REMOTE_READ,
-};
-
-static uint64_t smr_rx_comp_flags(uint32_t op, uint16_t op_flags)
-{
-	uint64_t flags;
-
-	flags = smr_rx_flags[op];
-
-	if (op_flags & OFI_REMOTE_CQ_DATA)
-		flags |= FI_REMOTE_CQ_DATA;
-
-	return flags;
-}
-
-static const uint64_t smr_mr_flags[] = {
-	[ofi_op_msg] = FI_RECV,
-	[ofi_op_tagged] = FI_RECV,
-	[ofi_op_read_req] = FI_REMOTE_READ,
-	[ofi_op_write] = FI_REMOTE_WRITE,
-	[ofi_op_atomic] = FI_REMOTE_WRITE,
-	[ofi_op_atomic_fetch] =  FI_REMOTE_WRITE | FI_REMOTE_READ,
-	[ofi_op_atomic_compare] = FI_REMOTE_WRITE | FI_REMOTE_READ,
-};
-
-static uint64_t smr_mr_reg_flags(uint32_t op, uint16_t atomic_op)
-{
-	if (atomic_op == FI_ATOMIC_READ)
-		return FI_REMOTE_READ;
-
-	return smr_mr_flags[op];
-}
-
-void smr_progress_resp(struct smr_ep *ep)
-{
-	struct smr_resp *resp;
-	struct dlist_entry *dlist_entry;
-	struct smr_pending_cmd *pending;
-	struct smr_match_attr match_attr;
-	int ret;
-
-	fastlock_acquire(&ep->region->lock);
-	fastlock_acquire(&ep->util_ep.tx_cq->cq_lock);
-	while (!ofi_cirque_isempty(smr_resp_queue(ep->region)) &&
-	       !ofi_cirque_isfull(ep->util_ep.tx_cq->cirq)) {
-		resp = ofi_cirque_head(smr_resp_queue(ep->region));
-		if (resp->status == FI_EBUSY)
-			break;
-		match_attr.ctx = resp->msg_id;
-		dlist_entry = dlist_remove_first_match(&ep->pend_queue.msg_list,
-						       ep->pend_queue.match_msg,
-						       &match_attr);
-		if (!dlist_entry) {
-			FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-				"no outstanding commands for found response\n");
-			break;
-		}
-		pending = container_of(dlist_entry, struct smr_pending_cmd, entry);
-
-		ret = ep->tx_comp(ep, (void *) resp->msg_id,
-				  smr_tx_comp_flags(pending->cmd.msg.hdr.op),
-				  -(resp->status));
-		if (ret) {
-			FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-				"unable to process tx completion\n");
-			break;
-		}
-		freestack_push(ep->pend_fs, pending);
-		ofi_cirque_discard(smr_resp_queue(ep->region));
-	}
-	fastlock_release(&ep->util_ep.tx_cq->cq_lock);
-	fastlock_release(&ep->region->lock);
-}
-
-static int smr_progress_inline(struct smr_cmd *cmd, struct iovec *iov,
-			       size_t iov_count, size_t *total_len)
-{
-	*total_len = ofi_copy_to_iov(iov, iov_count, 0, cmd->msg.data.msg,
-				     cmd->msg.hdr.size);
-	if (*total_len != cmd->msg.hdr.size) {
-		FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-			"recv truncated");
-		return -FI_EIO;
-	}
-	return 0;
-}
-
-static int smr_progress_inject(struct smr_cmd *cmd, struct iovec *iov,
-			       size_t iov_count, size_t *total_len,
-			       struct smr_ep *ep, int err)
-{
-	struct smr_inject_buf *tx_buf;
-	size_t inj_offset;
-
-	inj_offset = (size_t) cmd->msg.hdr.src_data;
-	tx_buf = (struct smr_inject_buf *) ((char **) ep->region +
-					    inj_offset);
-	if (err)
-		goto out;
-
-	*total_len = ofi_copy_to_iov(iov, iov_count, 0, tx_buf->data,
-				     cmd->msg.hdr.size);
-	if (*total_len != cmd->msg.hdr.size) {
-		FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-			"recv truncated");
-		err = -FI_EIO;
-	}
-
-out:
-	smr_freestack_push(smr_inject_pool(ep->region), tx_buf);
-	return err;
-}
-
-static int smr_progress_iov(struct smr_cmd *cmd, struct iovec *iov,
-			    size_t iov_count, size_t *total_len,
-			    struct smr_ep *ep, int err)
-{
-	struct smr_region *peer_smr;
-	struct smr_resp *resp;
-	int peer_id, ret;
-
-	peer_id = (int) cmd->msg.hdr.addr;
-	peer_smr = smr_peer_region(ep->region, peer_id);
-	resp = (struct smr_resp *) ((char **) peer_smr +
-				    (size_t) cmd->msg.hdr.src_data);
-
-	if (err) {
-		ret = -err;
-		goto out;
-	}
-
-	if (cmd->msg.hdr.op == ofi_op_read_req) {
-		ret = process_vm_writev(peer_smr->pid, iov, iov_count,
-					cmd->msg.data.iov,
-					cmd->msg.data.iov_count, 0);
-	} else {
-		ret = process_vm_readv(peer_smr->pid, iov, iov_count,
-				       cmd->msg.data.iov,
-				       cmd->msg.data.iov_count, 0);
-	}
-
-	if (ret != cmd->msg.hdr.size) {
-		if (ret < 0) {
-			FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-				"CMA write error\n");
-			ret = errno;
-		} else { 
-			FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-				"partial read occurred\n");
-			ret = FI_EIO;
-		}
-	} else {
-		*total_len = ret;
-		ret = 0;
-	}
-
-out:
-	//Status must be set last (signals peer: op done, valid resp entry)
-	resp->status = ret;
-
-	return -ret;
-}
-
-static void smr_do_atomic(void *src, void *dst, void *cmp, enum fi_datatype datatype,
-			  enum fi_op op, size_t cnt)
-{
-	char tmp_result[SMR_INJECT_SIZE];
-
-	if (op >= OFI_SWAP_OP_START) {
-		ofi_atomic_swap_handlers[op - OFI_SWAP_OP_START][datatype](dst,
-			src, cmp, tmp_result, cnt);
-	} else if (op != FI_ATOMIC_READ) {
-		ofi_atomic_write_handlers[op][datatype](dst, src, cnt);
-	}
-}
-
-static int smr_progress_inline_atomic(struct smr_cmd *cmd, struct fi_ioc *ioc,
-			       size_t ioc_count, size_t *len)
-{
-	int i;
-	uint8_t *src, *comp;
-
-
-	switch (cmd->msg.hdr.op) {
-	case ofi_op_atomic:
-	case ofi_op_atomic_fetch:
-		src = cmd->msg.data.msg;
-		break;
-	case ofi_op_atomic_compare:
-		src = cmd->msg.data.buf;
-		comp = cmd->msg.data.comp;
-		break;
-	default:
-		break;
-	}
-
-	for (i = *len = 0; i < ioc_count && *len < cmd->msg.hdr.size; i++) {
-		smr_do_atomic(&src[*len], ioc[i].addr, &comp[*len],
-			      cmd->msg.hdr.datatype, cmd->msg.hdr.atomic_op,
-			      ioc[i].count);
-		*len += ioc[i].count * ofi_datatype_size(cmd->msg.hdr.datatype);
-	}
-
-	if (*len != cmd->msg.hdr.size) {
-		FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-			"recv truncated");
-		return -FI_EIO;
-	}
-	return 0;
-}
-
-static int smr_progress_inject_atomic(struct smr_cmd *cmd, struct fi_ioc *ioc,
-			       size_t ioc_count, size_t *len,
-			       struct smr_ep *ep, int err)
-{
-	struct smr_inject_buf *tx_buf;
-	size_t inj_offset;
-	uint8_t *src, *comp;
-	int i;
-
-	inj_offset = (size_t) cmd->msg.hdr.src_data;
-	tx_buf = (struct smr_inject_buf *) ((char **) ep->region +
-					    inj_offset);
-	if (err)
-		goto out;
-
-	if (cmd->msg.hdr.op == ofi_op_atomic_compare) {
-		src = tx_buf->buf;
-		comp = tx_buf->comp;
-	} else {
-		src = tx_buf->data;
-	}
-
-	for (i = *len = 0; i < ioc_count && *len < cmd->msg.hdr.size; i++) {
-		smr_do_atomic(&src[*len], ioc[i].addr, &comp[*len],
-			      cmd->msg.hdr.datatype, cmd->msg.hdr.atomic_op,
-			      ioc[i].count);
-		*len += ioc[i].count * ofi_datatype_size(cmd->msg.hdr.datatype);
-	}
-
-	if (*len != cmd->msg.hdr.size) {
-		FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-			"recv truncated");
-		err = -FI_EIO;
-	}
-
-out:
-	smr_freestack_push(smr_inject_pool(ep->region), tx_buf);
-	return err;
-}
-
-static int smr_progress_cmd_msg(struct smr_ep *ep, struct smr_cmd *cmd)
-{
-	struct smr_recv_queue *recv_queue;
-	struct smr_match_attr match_attr;
-	struct dlist_entry *dlist_entry;
-	struct smr_ep_entry *entry;
-	struct smr_pending_cmd *unexp;
-	fi_addr_t addr;
-	size_t total_len = 0;
-	int err, ret = 0;
-
-	if (ofi_cirque_isfull(ep->util_ep.rx_cq->cirq)) {
-		FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-			"rx cq full\n");
-		return -FI_ENOSPC;
-	}
-
-	recv_queue = (cmd->msg.hdr.op == ofi_op_tagged) ?
-		      &ep->trecv_queue : &ep->recv_queue;
-
-	if (dlist_empty(&recv_queue->recv_list)) {
-		FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-			"no recv entry available\n");
-		return -FI_ENOMSG;
-	}
-
-	match_attr.addr = cmd->msg.hdr.addr;
-	match_attr.tag = cmd->msg.hdr.tag;
-
-	dlist_entry = dlist_remove_first_match(&recv_queue->recv_list,
-					       recv_queue->match_recv,
-					       &match_attr);
-	if (!dlist_entry) {
-		if (freestack_isempty(ep->unexp_fs))
-			return -FI_EAGAIN;
-		unexp = freestack_pop(ep->unexp_fs);
-		memcpy(&unexp->cmd, cmd, sizeof(*cmd));
-		ofi_cirque_discard(smr_cmd_queue(ep->region));
-		dlist_insert_tail(&unexp->entry, &ep->unexp_queue.msg_list);
-		return ret;
-	}
-	entry = container_of(dlist_entry, struct smr_ep_entry, entry);
-
-	switch (cmd->msg.hdr.op_src) {
-	case smr_src_inline:
-		err = smr_progress_inline(cmd, entry->iov, entry->iov_count,
-					  &total_len);
-		break;
-	case smr_src_inject:
-		err = smr_progress_inject(cmd, entry->iov, entry->iov_count,
-					  &total_len, ep, 0);
-		break;
-	case smr_src_iov:
-		err = smr_progress_iov(cmd, entry->iov, entry->iov_count,
-				       &total_len, ep, 0);
-		break;
-	default:
-		FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-			"unidentified operation type\n");
-		err = -FI_EINVAL;
-	}
-	ret = ep->rx_comp(ep, entry->context, smr_rx_comp_flags(cmd->msg.hdr.op,
-			  cmd->msg.hdr.op_flags), total_len,
-			  entry->iov[0].iov_base, &addr, cmd->msg.hdr.tag,
-			  cmd->msg.hdr.data, err);
-	if (ret) {
-		FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-			"unable to process rx completion\n");
-	}
-	freestack_push(ep->recv_fs, entry);
-	ofi_cirque_discard(smr_cmd_queue(ep->region));
-	ep->region->cmd_cnt++;
-
-	return ret;
-}
-
-static int smr_progress_cmd_rma(struct smr_ep *ep, struct smr_cmd *cmd)
-{
-	struct smr_domain *domain;
-	struct smr_cmd *rma_cmd;
-	struct iovec iov[SMR_IOV_LIMIT];
-	size_t iov_count;
-	size_t total_len = 0;
-	int err, ret = 0;
-
-	domain = container_of(ep->util_ep.domain, struct smr_domain,
-			      util_domain);
-
-	if (cmd->msg.hdr.op_flags & OFI_REMOTE_CQ_DATA &&
-	    ofi_cirque_isfull(ep->util_ep.rx_cq->cirq)) {
-		FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-			"rx cq full\n");
-		return -FI_ENOSPC;
-	}
-
-	ofi_cirque_discard(smr_cmd_queue(ep->region));
-	ep->region->cmd_cnt++;
-	rma_cmd = ofi_cirque_head(smr_cmd_queue(ep->region));
-
-	for (iov_count = 0; iov_count < rma_cmd->rma.rma_count; iov_count++) {
-		ret = ofi_mr_verify(&domain->util_domain.mr_map,
-				rma_cmd->rma.rma_iov[iov_count].len,
-				(uintptr_t *) &(rma_cmd->rma.rma_iov[iov_count].addr),
-				rma_cmd->rma.rma_iov[iov_count].key,
-				smr_mr_reg_flags(cmd->msg.hdr.op, 0));
-		if (ret)
-			break;
-
-		iov[iov_count].iov_base = (void *) rma_cmd->rma.rma_iov[iov_count].addr;
-		iov[iov_count].iov_len = rma_cmd->rma.rma_iov[iov_count].len;
-	}
-	ofi_cirque_discard(smr_cmd_queue(ep->region));
-	ep->region->cmd_cnt++;
-	if (ret)
-		return ret;
-
-	switch (cmd->msg.hdr.op_src) {
-	case smr_src_inline:
-		err = smr_progress_inline(cmd, iov, iov_count, &total_len);
-		break;
-	case smr_src_inject:
-		err = smr_progress_inject(cmd, iov, iov_count, &total_len, ep, ret);
-		break;
-	case smr_src_iov:
-		err = smr_progress_iov(cmd, iov, iov_count, &total_len, ep, ret);
-		break;
-	default:
-		FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-			"unidentified operation type\n");
-		err = -FI_EINVAL;
-	}
-	if (cmd->msg.hdr.op_flags & OFI_REMOTE_CQ_DATA) {
-		ret = ep->rx_comp(ep, (void *) cmd->msg.hdr.msg_id,
-				  smr_rx_comp_flags(cmd->msg.hdr.op,
-				  cmd->msg.hdr.op_flags), total_len,
-				  NULL, &cmd->msg.hdr.addr, 0,
-				  cmd->msg.hdr.data, err);
-		if (ret) {
-			FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-				"unable to process rx completion\n");
-		}
-	}
-
-	return ret;
-}
-
-static int smr_progress_cmd_atomic(struct smr_ep *ep, struct smr_cmd *cmd)
-{
-	struct smr_domain *domain;
-	struct smr_cmd *rma_cmd;
-	struct fi_ioc ioc[SMR_IOV_LIMIT];
-	size_t ioc_count;
-	size_t total_len = 0;
-	int err, ret = 0;
-
-	domain = container_of(ep->util_ep.domain, struct smr_domain,
-			      util_domain);
-
-	ofi_cirque_discard(smr_cmd_queue(ep->region));
-	ep->region->cmd_cnt++;
-	rma_cmd = ofi_cirque_head(smr_cmd_queue(ep->region));
-
-	for (ioc_count = 0; ioc_count < rma_cmd->rma.rma_count; ioc_count++) {
-		ret = ofi_mr_verify(&domain->util_domain.mr_map,
-				rma_cmd->rma.rma_ioc[ioc_count].count *
-				ofi_datatype_size(cmd->msg.hdr.datatype),
-				(uintptr_t *) &(rma_cmd->rma.rma_ioc[ioc_count].addr),
-				rma_cmd->rma.rma_ioc[ioc_count].key,
-				smr_mr_reg_flags(cmd->msg.hdr.op,
-				cmd->msg.hdr.atomic_op));
-		if (ret)
-			break;
-
-		ioc[ioc_count].addr = (void *) rma_cmd->rma.rma_ioc[ioc_count].addr;
-		ioc[ioc_count].count = rma_cmd->rma.rma_ioc[ioc_count].count;
-	}
-	ofi_cirque_discard(smr_cmd_queue(ep->region));
-	ep->region->cmd_cnt++;
-	if (ret)
-		return ret;
-
-	switch (cmd->msg.hdr.op_src) {
-	case smr_src_inline:
-		err = smr_progress_inline_atomic(cmd, ioc, ioc_count, &total_len);
-		break;
-	case smr_src_inject:
-		err = smr_progress_inject_atomic(cmd, ioc, ioc_count, &total_len, ep, ret);
-		break;
-	default:
-		FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-			"unidentified operation type\n");
-		err = -FI_EINVAL;
-	}
-
-	if (err)
-		FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-			"error processing atomic op\n");
-
-	return err; 
-}
-
-void smr_progress_cmd(struct smr_ep *ep)
-{
-	struct smr_cmd *cmd;
-	int ret = 0;
-
-	fastlock_acquire(&ep->region->lock);
-	fastlock_acquire(&ep->util_ep.rx_cq->cq_lock);
-
-	while (!ofi_cirque_isempty(smr_cmd_queue(ep->region))) {
-		cmd = ofi_cirque_head(smr_cmd_queue(ep->region));
-
-		switch (cmd->msg.hdr.op) {
-		case ofi_op_msg:
-		case ofi_op_tagged:
-			ret = smr_progress_cmd_msg(ep, cmd);
-			break;
-		case ofi_op_write:
-		case ofi_op_read_req:
-			ret = smr_progress_cmd_rma(ep, cmd);
-			break;
-		case ofi_op_write_rsp:
-		case ofi_op_read_rsp:
-			ofi_cirque_discard(smr_cmd_queue(ep->region));
-			ep->region->cmd_cnt++;
-			break;
-		case ofi_op_atomic:
-		case ofi_op_atomic_fetch:
-		case ofi_op_atomic_compare:
-			ret = smr_progress_cmd_atomic(ep, cmd);
-			break;
-		default:
-			FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-				"unidentified operation type\n");
-			ret = -FI_EINVAL;
-		}
-
-		if (ret) {
-			if (ret != -FI_EAGAIN) {
-				FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-					"error processing command\n");
-			}
-			break;
-		}
-	}
-	fastlock_release(&ep->util_ep.rx_cq->cq_lock);
-	fastlock_release(&ep->region->lock);
 }
 
 static int smr_match_msg(struct dlist_entry *item, const void *args)
@@ -798,153 +174,7 @@ static void smr_init_pending_queue(struct smr_pending_queue *queue,
 	queue->match_msg = match_func;
 }
 
-void smr_ep_progress(struct util_ep *util_ep)
-{
-	struct smr_ep *ep;
-
-	ep = container_of(util_ep, struct smr_ep, util_ep);
-
-	smr_progress_resp(ep);
-	smr_progress_cmd(ep);
-}
-
-static int smr_check_unexp(struct smr_ep *ep, struct smr_ep_entry *entry)
-{
-	struct smr_match_attr match_attr;
-	struct smr_pending_cmd *unexp_msg;
-	struct dlist_entry *dlist_entry;
-	size_t total_len = 0;
-	int ret = 0;
-
-	if (ofi_cirque_isfull(ep->util_ep.rx_cq->cirq)) {
-		FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-			"rx cq full\n");
-		ret = -FI_EAGAIN;
-		goto push_entry;
-	}
-
-	match_attr.addr = entry->addr;
-	match_attr.ignore = entry->ignore;
-	match_attr.tag = entry->tag;
-	dlist_entry = dlist_remove_first_match(&ep->unexp_queue.msg_list,
-					       ep->unexp_queue.match_msg,
-					       &match_attr);
-	if (!dlist_entry)
-		return -FI_ENOMSG;
-
-	unexp_msg = container_of(dlist_entry, struct smr_pending_cmd, entry);
-
-	switch (unexp_msg->cmd.msg.hdr.op_src) {
-	case smr_src_inline:
-		entry->err = smr_progress_inline(&unexp_msg->cmd, entry->iov,
-						 entry->iov_count, &total_len);
-		break;
-	case smr_src_inject:
-		entry->err = smr_progress_inject(&unexp_msg->cmd, entry->iov,
-						 entry->iov_count, &total_len,
-						 ep, 0);
-		break;
-	case smr_src_iov:
-		entry->err = smr_progress_iov(&unexp_msg->cmd, entry->iov,
-					      entry->iov_count, &total_len,
-					      ep, 0);
-		break;
-	default:
-		FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-			"unidentified operation type\n");
-		entry->err = FI_EINVAL;
-	}
-	ret = ep->rx_comp(ep, entry->context,
-			  smr_rx_comp_flags(unexp_msg->cmd.msg.hdr.op,
-			  unexp_msg->cmd.msg.hdr.op_flags), total_len,
-			  entry->iov[0].iov_base, &entry->addr, entry->tag,
-			  unexp_msg->cmd.msg. hdr.data, entry->err);
-	if (ret) {
-		FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-			"unable to process rx completion\n");
-	}
-
-	ep->region->cmd_cnt++;
-	freestack_push(ep->unexp_fs, unexp_msg);
-push_entry:
-	freestack_push(ep->recv_fs, entry);
-	return ret;
-}
-
-static ssize_t smr_generic_recvmsg(struct fid_ep *ep_fid, const struct iovec *iov,
-				   size_t iov_count, fi_addr_t addr, uint64_t tag,
-				   uint64_t ignore, void *context, uint64_t flags)
-{
-	struct smr_ep *ep;
-	struct smr_recv_queue *recv_queue;
-	struct smr_ep_entry *entry;
-	ssize_t ret;
-
-	assert(iov_count <= SMR_IOV_LIMIT);
-
-	ep = container_of(ep_fid, struct smr_ep, util_ep.ep_fid.fid);
-	fastlock_acquire(&ep->util_ep.rx_cq->cq_lock);
-	if (freestack_isempty(ep->recv_fs)) {
-		ret = -FI_EAGAIN;
-		goto out;
-	}
-	entry = freestack_pop(ep->recv_fs);
-	memset(entry, 0, sizeof(*entry));
-
-	for (entry->iov_count = 0; entry->iov_count < iov_count;
-	     entry->iov_count++) {
-		entry->iov[entry->iov_count] = iov[entry->iov_count];
-	}
-
-	entry->context = context;
-	entry->flags = flags;
-	entry->addr = addr;
-	entry->tag = tag;
-	entry->ignore = ignore;
-
-	if (flags & FI_TAGGED) {
-		ret = smr_check_unexp(ep, entry);
-		if (!ret || ret == -FI_EAGAIN)
-			goto out;
-		recv_queue = &ep->trecv_queue;
-	} else {
-		recv_queue = &ep->recv_queue;
-	}
-
-	dlist_insert_tail(&entry->entry, &recv_queue->recv_list);
-
-	ret = 0;
-out:
-	fastlock_release(&ep->util_ep.rx_cq->cq_lock);
-	return ret;
-}
-
-ssize_t smr_recvmsg(struct fid_ep *ep_fid, const struct fi_msg *msg,
-		    uint64_t flags)
-{
-	return smr_generic_recvmsg(ep_fid, msg->msg_iov, msg->iov_count,
-				   msg->addr, 0, 0, msg->context, flags);
-}
-
-ssize_t smr_recvv(struct fid_ep *ep_fid, const struct iovec *iov, void **desc,
-		size_t count, fi_addr_t src_addr, void *context)
-{
-	return smr_generic_recvmsg(ep_fid, iov, count, src_addr,
-				   0, 0, context, 0);
-}
-
-ssize_t smr_recv(struct fid_ep *ep_fid, void *buf, size_t len, void *desc,
-		fi_addr_t src_addr, void *context)
-{
-	struct iovec msg_iov;
-
-	msg_iov.iov_base = (void *) buf;
-	msg_iov.iov_len = len;
-
-	return smr_generic_recvmsg(ep_fid, &msg_iov, 1, src_addr, 0, 0, context, 0);
-}
-
-static void smr_post_pending(struct smr_ep *ep, struct smr_cmd *cmd)
+void smr_post_pending(struct smr_ep *ep, struct smr_cmd *cmd)
 {
 	struct smr_pending_cmd *pend_cmd;
 
@@ -954,10 +184,10 @@ static void smr_post_pending(struct smr_ep *ep, struct smr_cmd *cmd)
 	dlist_insert_tail(&pend_cmd->entry, &ep->pend_queue.msg_list);
 }
 
-static void smr_generic_format(struct smr_cmd *cmd, fi_addr_t peer_id, void *context,
-			       uint32_t op, uint64_t tag, uint8_t datatype,
-			       uint8_t atomic_op, uint64_t data,
-			       uint16_t op_flags)
+void smr_generic_format(struct smr_cmd *cmd, fi_addr_t peer_id, void *context,
+			uint32_t op, uint64_t tag, uint8_t datatype,
+			uint8_t atomic_op, uint64_t data,
+			uint16_t op_flags)
 {
 	cmd->msg.hdr.op = op;
 	cmd->msg.hdr.op_flags = op_flags;
@@ -974,24 +204,10 @@ static void smr_generic_format(struct smr_cmd *cmd, fi_addr_t peer_id, void *con
 	cmd->msg.hdr.data = data;
 }
 
-static void smr_format_rma_iov(struct smr_cmd *cmd, const struct fi_rma_iov *rma_iov,
-			       size_t iov_count)
-{
-	cmd->rma.rma_count = iov_count;
-	memcpy(cmd->rma.rma_iov, rma_iov, sizeof(*rma_iov) * iov_count);
-}
-
-static void smr_format_rma_ioc(struct smr_cmd *cmd, const struct fi_rma_ioc *rma_ioc,
-			       size_t ioc_count)
-{
-	cmd->rma.rma_count = ioc_count;
-	memcpy(cmd->rma.rma_ioc, rma_ioc, sizeof(*rma_ioc) * ioc_count);
-}
-
-static void smr_format_inline(struct smr_cmd *cmd, fi_addr_t peer_id,
-			      const struct iovec *iov, size_t count,
-			      void *context, uint32_t op, uint64_t tag,
-			      uint64_t data, uint16_t op_flags)
+void smr_format_inline(struct smr_cmd *cmd, fi_addr_t peer_id,
+		       const struct iovec *iov, size_t count,
+		       void *context, uint32_t op, uint64_t tag,
+		       uint64_t data, uint16_t op_flags)
 {
 	smr_generic_format(cmd, peer_id, context, op, tag, 0, 0, data,
 			   op_flags);
@@ -1000,12 +216,12 @@ static void smr_format_inline(struct smr_cmd *cmd, fi_addr_t peer_id,
 					      SMR_MSG_DATA_LEN, iov, count, 0);
 }
 
-static void smr_format_inject(struct smr_cmd *cmd, fi_addr_t peer_id,
-			      const struct iovec *iov, size_t count,
-			      void *context, uint32_t op, uint64_t tag,
-			      uint64_t data, uint16_t op_flags,
-			      struct smr_region *smr,
-			      struct smr_inject_buf *tx_buf)
+void smr_format_inject(struct smr_cmd *cmd, fi_addr_t peer_id,
+		       const struct iovec *iov, size_t count,
+		       void *context, uint32_t op, uint64_t tag,
+		       uint64_t data, uint16_t op_flags,
+		       struct smr_region *smr,
+		       struct smr_inject_buf *tx_buf)
 {
 	smr_generic_format(cmd, peer_id, context, op, tag, 0, 0,
 			   data, op_flags);
@@ -1015,81 +231,12 @@ static void smr_format_inject(struct smr_cmd *cmd, fi_addr_t peer_id,
 					       iov, count, 0);
 }
 
-static void smr_format_inline_atomic(struct smr_cmd *cmd, fi_addr_t peer_id,
-				     const struct iovec *iov, size_t count,
-				     const struct iovec *compv,
-				     size_t comp_count, void *context,
-				     uint32_t op, enum fi_datatype datatype,
-				     enum fi_op atomic_op)
-{
-	size_t comp_size;
-
-	smr_generic_format(cmd, peer_id, context, op, 0, datatype,
-			   atomic_op, 0, 0);
-	cmd->msg.hdr.op_src = smr_src_inline;
-	switch (op) {
-	case ofi_op_atomic:
-	case ofi_op_atomic_fetch:
-		cmd->msg.hdr.size = ofi_copy_from_iov(cmd->msg.data.msg,
-						SMR_MSG_DATA_LEN, iov, count, 0);
-		break;
-	case ofi_op_atomic_compare:
-		cmd->msg.hdr.size = ofi_copy_from_iov(cmd->msg.data.buf,
-						SMR_MSG_DATA_LEN, iov, count, 0);
-		comp_size = ofi_copy_from_iov(cmd->msg.data.comp,
-					      SMR_MSG_DATA_LEN, compv,
-					      comp_count, 0);
-		if (comp_size != cmd->msg.hdr.size)
-			FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-				"atomic and compare buffer size mimatch\n");
-		break;
-	default:
-		break;
-	}
-}
-
-static void smr_format_inject_atomic(struct smr_cmd *cmd, fi_addr_t peer_id,
-				     const struct iovec *iov, size_t count,
-				     const struct iovec *compv,
-				     size_t comp_count,void *context,
-				     uint32_t op, enum fi_datatype datatype,
-				     enum fi_op atomic_op,
-				     struct smr_region *smr,
-				     struct smr_inject_buf *tx_buf)
-{
-	size_t comp_size;
-
-	smr_generic_format(cmd, peer_id, context, op, 0, datatype,
-			   atomic_op, 0, 0);
-	cmd->msg.hdr.op_src = smr_src_inject;
-	cmd->msg.hdr.src_data = (char **) tx_buf - (char **) smr;
-
-	switch (op) {
-	case ofi_op_atomic:
-	case ofi_op_atomic_fetch:
-		cmd->msg.hdr.size = ofi_copy_from_iov(tx_buf->data,
-						SMR_INJECT_SIZE, iov, count, 0);
-		break;
-	case ofi_op_atomic_compare:
-		cmd->msg.hdr.size = ofi_copy_from_iov(tx_buf->buf,
-						SMR_COMP_INJECT_SIZE, iov, count, 0);
-		comp_size = ofi_copy_from_iov(tx_buf->comp, SMR_COMP_INJECT_SIZE,
-					      compv, comp_count, 0);
-		if (comp_size != cmd->msg.hdr.size)
-			FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-				"atomic and compare buffer size mimatch\n");
-		break;
-	default:
-		break;
-	}
-}
-
-static void smr_format_iov(struct smr_cmd *cmd, fi_addr_t peer_id,
-			   const struct iovec *iov, size_t count,
-			   size_t total_len, void *context, uint32_t op,
-			   uint64_t tag, uint64_t data, uint16_t op_flags,
-			   struct smr_region *smr,
-			   struct smr_resp *resp)
+void smr_format_iov(struct smr_cmd *cmd, fi_addr_t peer_id,
+		    const struct iovec *iov, size_t count,
+		    size_t total_len, void *context, uint32_t op,
+		    uint64_t tag, uint64_t data, uint16_t op_flags,
+		    struct smr_region *smr,
+		    struct smr_resp *resp)
 {
 	smr_generic_format(cmd, peer_id, context, op, tag, 0, 0, data, op_flags);
 	cmd->msg.hdr.op_src = smr_src_iov;
@@ -1101,1173 +248,6 @@ static void smr_format_iov(struct smr_cmd *cmd, fi_addr_t peer_id,
 
 	memcpy(cmd->msg.data.iov, iov, sizeof(*iov) * count);
 }
-
-static void smr_format_rma_resp(struct smr_cmd *cmd, fi_addr_t peer_id,
-				const struct fi_rma_iov *rma_iov, size_t count,
-				size_t total_len, uint32_t op)
-{
-	smr_generic_format(cmd, peer_id, NULL, op, 0, 0, 0, 0, 0);
-	cmd->msg.hdr.size = total_len;
-}
-
-static ssize_t smr_generic_sendmsg(struct fid_ep *ep_fid, const struct iovec *iov,
-				   size_t iov_count, fi_addr_t addr, uint64_t tag,
-				   uint64_t data, void *context, uint32_t op,
-				   uint16_t op_flags)
-{
-	struct smr_ep *ep;
-	struct smr_region *peer_smr;
-	struct smr_inject_buf *tx_buf;
-	struct smr_resp *resp;
-	struct smr_cmd *cmd;
-	int peer_id;
-	ssize_t ret = 0;
-	size_t total_len;
-
-	assert(iov_count <= SMR_IOV_LIMIT);
-
-	ep = container_of(ep_fid, struct smr_ep, util_ep.ep_fid.fid);
-	peer_id = (int) addr;
-
-	ret = smr_verify_peer(ep, peer_id);
-	if (ret)
-		return ret;
-
-	peer_smr = smr_peer_region(ep->region, peer_id);
-	fastlock_acquire(&peer_smr->lock);
-	if (!peer_smr->cmd_cnt) {
-		ret = -FI_EAGAIN;
-		goto unlock_region;
-	}
-
-	fastlock_acquire(&ep->util_ep.tx_cq->cq_lock);
-	if (ofi_cirque_isfull(ep->util_ep.tx_cq->cirq)) {
-		ret = -FI_EAGAIN;
-		goto unlock_cq;
-	}
-
-	total_len = ofi_total_iov_len(iov, iov_count);
-
-	cmd = ofi_cirque_tail(smr_cmd_queue(peer_smr));
-
-	if (total_len <= SMR_MSG_DATA_LEN) {
-		smr_format_inline(cmd, smr_peer_addr(ep->region)[peer_id].addr, iov,
-				  iov_count, context, op, tag, data, op_flags);
-	} else if (total_len <= SMR_INJECT_SIZE) {
-		tx_buf = smr_freestack_pop(smr_inject_pool(peer_smr));
-		smr_format_inject(cmd, smr_peer_addr(ep->region)[peer_id].addr, iov,
-				  iov_count, context, op, tag, data,
-				  op_flags, peer_smr, tx_buf);
-	} else {
-		assert(!ofi_cirque_isfull(smr_resp_queue(ep->region)));
-		resp = ofi_cirque_tail(smr_resp_queue(ep->region));
-		smr_format_iov(cmd, smr_peer_addr(ep->region)[peer_id].addr, iov,
-			       iov_count, total_len, context, op, tag, data,
-			       op_flags, ep->region, resp);
-		ofi_cirque_commit(smr_resp_queue(ep->region));
-		smr_post_pending(ep, cmd);
-		goto commit;
-	}
-	ret = ep->tx_comp(ep, context, smr_tx_comp_flags(op), 0);
-	if (ret) {
-		FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-			"unable to process tx completion\n");
-		goto unlock_cq;
-	}
-
-commit:
-	ofi_cirque_commit(smr_cmd_queue(peer_smr));
-	peer_smr->cmd_cnt--;
-unlock_cq:
-	fastlock_release(&ep->util_ep.tx_cq->cq_lock);
-unlock_region:
-	fastlock_release(&peer_smr->lock);
-	return ret;
-}
-
-ssize_t smr_send(struct fid_ep *ep_fid, const void *buf, size_t len, void *desc,
-		fi_addr_t dest_addr, void *context)
-{
-	struct iovec msg_iov;
-
-	msg_iov.iov_base = (void *) buf;
-	msg_iov.iov_len = len;
-
-	return smr_generic_sendmsg(ep_fid, &msg_iov, 1, dest_addr, 0,
-				   0, context, ofi_op_msg, 0);
-}
-
-
-ssize_t smr_sendv(struct fid_ep *ep_fid, const struct iovec *iov, void **desc,
-		size_t count, fi_addr_t dest_addr, void *context)
-{
-	return smr_generic_sendmsg(ep_fid, iov, count, dest_addr, 0,
-				   0, context, ofi_op_msg, 0);
-}
-
-ssize_t smr_sendmsg(struct fid_ep *ep_fid, const struct fi_msg *msg,
-		    uint64_t flags)
-{
-	return smr_generic_sendmsg(ep_fid, msg->msg_iov, msg->iov_count,
-				   msg->addr, 0, msg->data, msg->context,
-				   ofi_op_msg, (flags & FI_REMOTE_CQ_DATA) ?
-				   OFI_REMOTE_CQ_DATA : 0);
-}
-
-static ssize_t smr_generic_inject(struct fid_ep *ep_fid, const void *buf,
-				  size_t len, fi_addr_t dest_addr, uint64_t tag,
-				  uint64_t data, uint32_t op, uint16_t op_flags)
-{
-	struct smr_ep *ep;
-	struct smr_region *peer_smr;
-	struct smr_inject_buf *tx_buf;
-	struct smr_cmd *cmd;
-	int peer_id;
-	ssize_t ret = 0;
-	struct iovec msg_iov;
-
-	assert(len <= SMR_INJECT_SIZE);
-
-	msg_iov.iov_base = (void *) buf;
-	msg_iov.iov_len = len;
-
-	ep = container_of(ep_fid, struct smr_ep, util_ep.ep_fid.fid);
-	peer_id = (int) dest_addr;
-
-	ret = smr_verify_peer(ep, peer_id);
-	if (ret)
-		return ret;
-
-	peer_smr = smr_peer_region(ep->region, peer_id);
-	fastlock_acquire(&peer_smr->lock);
-	if (!peer_smr->cmd_cnt) {
-		ret = -FI_EAGAIN;
-		goto unlock;
-	}
-
-	cmd = ofi_cirque_tail(smr_cmd_queue(peer_smr));
-
-	if (len <= SMR_MSG_DATA_LEN) {
-		smr_format_inline(cmd, smr_peer_addr(ep->region)[peer_id].addr,
-				  &msg_iov, 1, NULL, op, tag, data, op_flags);
-	} else {
-		tx_buf = smr_freestack_pop(smr_inject_pool(peer_smr));
-		smr_format_inject(cmd, smr_peer_addr(ep->region)[peer_id].addr,
-				  &msg_iov, 1, NULL, op, tag, data,
-				  op_flags, peer_smr, tx_buf);
-	}
-
-	peer_smr->cmd_cnt--;
-	ofi_cirque_commit(smr_cmd_queue(peer_smr));
-unlock:
-	fastlock_release(&peer_smr->lock);
-
-	return ret;
-}
-
-ssize_t smr_inject(struct fid_ep *ep_fid, const void *buf, size_t len,
-		fi_addr_t dest_addr)
-{
-	return smr_generic_inject(ep_fid, buf, len, dest_addr, 0, 0,
-				  ofi_op_msg, 0);
-}
-
-ssize_t smr_senddata(struct fid_ep *ep_fid, const void *buf, size_t len,
-		     void *desc, uint64_t data, fi_addr_t dest_addr,
-		     void *context)
-{
-	struct iovec iov;
-
-	iov.iov_base = (void *) buf;
-	iov.iov_len = len;
-
-	return smr_generic_sendmsg(ep_fid, &iov, 1, dest_addr, 0, data, context,
-				   ofi_op_msg, OFI_REMOTE_CQ_DATA);
-}
-
-ssize_t smr_injectdata(struct fid_ep *ep_fid, const void *buf, size_t len,
-		       uint64_t data, fi_addr_t dest_addr)
-{
-	return smr_generic_inject(ep_fid, buf, len, dest_addr, 0, data,
-				  ofi_op_msg, OFI_REMOTE_CQ_DATA);
-}
-
-static struct fi_ops_msg smr_msg_ops = {
-	.size = sizeof(struct fi_ops_msg),
-	.recv = smr_recv,
-	.recvv = smr_recvv,
-	.recvmsg = smr_recvmsg,
-	.send = smr_send,
-	.sendv = smr_sendv,
-	.sendmsg = smr_sendmsg,
-	.inject = smr_inject,
-	.senddata = smr_senddata,
-	.injectdata = smr_injectdata,
-};
-
-ssize_t smr_trecv(struct fid_ep *ep_fid, void *buf, size_t len, void *desc,
-	fi_addr_t src_addr, uint64_t tag, uint64_t ignore, void *context)
-{
-	struct iovec msg_iov;
-
-	msg_iov.iov_base = (void *) buf;
-	msg_iov.iov_len = len;
-
-	return smr_generic_recvmsg(ep_fid, &msg_iov, 1, src_addr, tag, ignore,
-				   context, FI_TAGGED);
-}
-
-ssize_t smr_trecvv(struct fid_ep *ep_fid, const struct iovec *iov, void **desc,
-	size_t count, fi_addr_t src_addr, uint64_t tag, uint64_t ignore,
-	void *context)
-{
-	return smr_generic_recvmsg(ep_fid, iov, count, src_addr, tag, ignore,
-				   context, FI_TAGGED);
-}
-
-ssize_t smr_trecvmsg(struct fid_ep *ep_fid, const struct fi_msg_tagged *msg,
-	uint64_t flags)
-{
-	return smr_generic_recvmsg(ep_fid, msg->msg_iov, msg->iov_count, msg->addr,
-				   msg->tag, msg->ignore, msg->context,
-				   flags | FI_TAGGED);
-}
-
-ssize_t smr_tsend(struct fid_ep *ep_fid, const void *buf, size_t len,
-	void *desc, fi_addr_t dest_addr, uint64_t tag, void *context)
-{
-	struct iovec msg_iov;
-
-	msg_iov.iov_base = (void *) buf;
-	msg_iov.iov_len = len;
-
-	return smr_generic_sendmsg(ep_fid, &msg_iov, 1, dest_addr, tag,
-				   0, context, ofi_op_tagged, 0);
-}
-
-ssize_t smr_tsendv(struct fid_ep *ep_fid, const struct iovec *iov,
-	void **desc, size_t count, fi_addr_t dest_addr, uint64_t tag,
-	void *context)
-{
-	return smr_generic_sendmsg(ep_fid, iov, count, dest_addr, tag,
-				   0, context, ofi_op_tagged, 0);
-}
-
-ssize_t smr_tsendmsg(struct fid_ep *ep_fid, const struct fi_msg_tagged *msg,
-	uint64_t flags)
-{
-	return smr_generic_sendmsg(ep_fid, msg->msg_iov, msg->iov_count,
-				   msg->addr, msg->tag, msg->data, msg->context,
-				   ofi_op_tagged, (flags & FI_REMOTE_CQ_DATA) ?
-				   OFI_REMOTE_CQ_DATA : 0);
-}
-
-ssize_t smr_tinject(struct fid_ep *ep_fid, const void *buf, size_t len,
-		    fi_addr_t dest_addr, uint64_t tag)
-{
-	return smr_generic_inject(ep_fid, buf, len, dest_addr, tag, 0,
-				  ofi_op_tagged, 0);
-}
-
-ssize_t smr_tsenddata(struct fid_ep *ep_fid, const void *buf, size_t len,
-		      void *desc, uint64_t data, fi_addr_t dest_addr,
-		      uint64_t tag, void *context)
-{
-	struct iovec iov;
-
-	iov.iov_base = (void *) buf;
-	iov.iov_len = len;
-
-	return smr_generic_sendmsg(ep_fid, &iov, 1, dest_addr, tag, data, context,
-				   ofi_op_tagged, OFI_REMOTE_CQ_DATA);
-}
-
-ssize_t smr_tinjectdata(struct fid_ep *ep_fid, const void *buf, size_t len,
-			uint64_t data, fi_addr_t dest_addr, uint64_t tag)
-{
-	return smr_generic_inject(ep_fid, buf, len, dest_addr, tag, data,
-				  ofi_op_tagged, OFI_REMOTE_CQ_DATA);
-}
-
-static struct fi_ops_tagged smr_tagged_ops = {
-	.size = sizeof(struct fi_ops_tagged),
-	.recv = smr_trecv,
-	.recvv = smr_trecvv,
-	.recvmsg = smr_trecvmsg,
-	.send = smr_tsend,
-	.sendv = smr_tsendv,
-	.sendmsg = smr_tsendmsg,
-	.inject = smr_tinject,
-	.senddata = smr_tsenddata,
-	.injectdata = smr_tinjectdata,
-};
-
-ssize_t smr_rma_fast(struct smr_ep *ep, const struct iovec *iov,
-	size_t iov_count, const struct fi_rma_iov *rma_iov, size_t rma_count,
-	void **desc, int peer_id, void *context, uint32_t op)
-{
-	struct smr_region *peer_smr;
-	struct iovec rma_iovec[SMR_IOV_LIMIT];
-	struct smr_cmd *cmd;
-	size_t total_len;
-	int ret, i;
-
-	for (i = 0; i < rma_count; i++) {
-		rma_iovec[i].iov_base = (void *) rma_iov[i].addr;
-		rma_iovec[i].iov_len = rma_iov[i].len;
-	}
-
-	total_len = ofi_total_iov_len(iov, iov_count);
-
-	peer_smr = smr_peer_region(ep->region, peer_id);
-	fastlock_acquire(&peer_smr->lock);
-	if (!peer_smr->cmd_cnt) {
-		ret = -FI_EAGAIN;
-		goto unlock_region;
-	}
-
-	fastlock_acquire(&ep->util_ep.tx_cq->cq_lock);
-	if (ofi_cirque_isfull(ep->util_ep.tx_cq->cirq)) {
-		ret = -FI_EAGAIN;
-		goto unlock_cq;
-	}
-
-	if (op == ofi_op_write) {
-		ret = process_vm_writev(peer_smr->pid, iov, iov_count,
-					rma_iovec, rma_count, 0);
-	} else {
-		ret = process_vm_readv(peer_smr->pid, iov, iov_count,
-				       rma_iovec, rma_count, 0);
-	}
-
-	if (ret != total_len) {
-		if (ret < 0) {
-			FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-				"CMA write error\n");
-			ret = -errno;
-		} else {
-			FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-				"unable to process tx completion\n");
-			ret = -FI_EIO;
-		}
-		goto comp;
-	}
-	ret = 0;
-
-	cmd = ofi_cirque_tail(smr_cmd_queue(peer_smr));
-	smr_format_rma_resp(cmd, peer_id, rma_iov, rma_count, total_len,
-			    (op == ofi_op_write) ? ofi_op_write_rsp :
-			    ofi_op_read_rsp);
-
-	peer_smr->cmd_cnt--;
-	ofi_cirque_commit(smr_cmd_queue(peer_smr));
-comp:
-	ret = ep->tx_comp(ep, context, smr_tx_comp_flags(op), ret);
-unlock_cq:
-	fastlock_release(&ep->util_ep.tx_cq->cq_lock);
-unlock_region:
-	fastlock_release(&peer_smr->lock);
-	return ret;
-}
-
-ssize_t smr_generic_rma(struct fid_ep *ep_fid, const struct iovec *iov,
-	size_t iov_count, const struct fi_rma_iov *rma_iov, size_t rma_count,
-	void **desc, fi_addr_t addr, void *context, uint32_t op, uint64_t data,
-	uint16_t op_flags)
-{
-	struct smr_ep *ep;
-	struct smr_region *peer_smr;
-	struct smr_inject_buf *tx_buf;
-	struct smr_resp *resp;
-	struct smr_cmd *cmd;
-	int peer_id, comp = 1;
-	ssize_t ret = 0;
-	size_t total_len;
-
-	assert(iov_count <= SMR_IOV_LIMIT);
-	assert(rma_count <= SMR_IOV_LIMIT);
-
-	ep = container_of(ep_fid, struct smr_ep, util_ep.ep_fid.fid);
-
-	peer_id = (int) addr;
-	ret = smr_verify_peer(ep, peer_id);
-	if (ret)
-		return ret;
-
-	if (rma_count == 1 && !(op_flags & OFI_REMOTE_CQ_DATA))
-		return smr_rma_fast(ep, iov, iov_count, rma_iov, rma_count,
-				    desc, peer_id, context, op);
-
-	peer_smr = smr_peer_region(ep->region, peer_id);
-	fastlock_acquire(&peer_smr->lock);
-	if (peer_smr->cmd_cnt < 2) {
-		ret = -FI_EAGAIN;
-		goto unlock_region;
-	}
-
-	fastlock_acquire(&ep->util_ep.tx_cq->cq_lock);
-	if (ofi_cirque_isfull(ep->util_ep.tx_cq->cirq)) {
-		ret = -FI_EAGAIN;
-		goto unlock_cq;
-	}
-
-	cmd = ofi_cirque_tail(smr_cmd_queue(peer_smr));
-	total_len = ofi_total_iov_len(iov, iov_count);
-
-	if (total_len <= SMR_MSG_DATA_LEN && op == ofi_op_write) {
-		smr_format_inline(cmd, smr_peer_addr(ep->region)[peer_id].addr,
-				  iov, iov_count, context, op, 0, data, op_flags);
-	} else if (total_len <= SMR_INJECT_SIZE && op == ofi_op_write) {
-		tx_buf = smr_freestack_pop(smr_inject_pool(peer_smr));
-		smr_format_inject(cmd, smr_peer_addr(ep->region)[peer_id].addr, iov,
-				  iov_count, context, op, 0, data, op_flags,
-				  peer_smr, tx_buf);
-	} else {
-		assert(!ofi_cirque_isfull(smr_resp_queue(ep->region)));
-		resp = ofi_cirque_tail(smr_resp_queue(ep->region));
-		smr_format_iov(cmd, smr_peer_addr(ep->region)[peer_id].addr, iov,
-			       iov_count, total_len, context, op, 0, data, op_flags,
-			       ep->region, resp);
-		ofi_cirque_commit(smr_resp_queue(ep->region));
-		smr_post_pending(ep, cmd);
-		comp = 0;
-	}
-
-	ofi_cirque_commit(smr_cmd_queue(peer_smr));
-	peer_smr->cmd_cnt--;
-	cmd = ofi_cirque_tail(smr_cmd_queue(peer_smr));
-	smr_format_rma_iov(cmd, rma_iov, rma_count);
-	ofi_cirque_commit(smr_cmd_queue(peer_smr));
-	peer_smr->cmd_cnt--;
-
-	if (!comp)
-		goto unlock_cq;
-
-	ret = ep->tx_comp(ep, context, smr_tx_comp_flags(op), 0);
-	if (ret) {
-		FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-			"unable to process tx completion\n");
-	}
-
-unlock_cq:
-	fastlock_release(&ep->util_ep.tx_cq->cq_lock);
-unlock_region:
-	fastlock_release(&peer_smr->lock);
-	return ret;
-}
-
-ssize_t smr_read(struct fid_ep *ep_fid, void *buf, size_t len, void *desc,
-	fi_addr_t src_addr, uint64_t addr, uint64_t key, void *context)
-{
-	struct iovec msg_iov;
-	struct fi_rma_iov rma_iov;
-
-	msg_iov.iov_base = (void *) buf;
-	msg_iov.iov_len = len;
-	rma_iov.addr = addr;
-	rma_iov.len = len;
-	rma_iov.key = key;
-
-	return smr_generic_rma(ep_fid, &msg_iov, 1, &rma_iov, 1, &desc, 
-			       src_addr, context, ofi_op_read_req, 0, 0);
-}
-
-ssize_t smr_readv(struct fid_ep *ep_fid, const struct iovec *iov, void **desc,
-	size_t count, fi_addr_t src_addr, uint64_t addr, uint64_t key,
-	void *context)
-{
-	struct fi_rma_iov rma_iov;
-
-	rma_iov.addr = addr;
-	rma_iov.len  = ofi_total_iov_len(iov, count);
-	rma_iov.key = key;
-
-	return smr_generic_rma(ep_fid, iov, count, &rma_iov, 1, desc,
-			       src_addr, context, ofi_op_read_req, 0, 0);
-}
-
-ssize_t smr_readmsg(struct fid_ep *ep_fid, const struct fi_msg_rma *msg,
-	uint64_t flags)
-{
-	return smr_generic_rma(ep_fid, msg->msg_iov, msg->iov_count,
-			       msg->rma_iov, msg->rma_iov_count,
-			       msg->desc, msg->addr, msg->context,
-			       ofi_op_read_req, 0, 0);
-}
-
-ssize_t smr_write(struct fid_ep *ep_fid, const void *buf, size_t len, void *desc,
-	fi_addr_t dest_addr, uint64_t addr, uint64_t key, void *context)
-{
-	struct iovec msg_iov;
-	struct fi_rma_iov rma_iov;
-
-	msg_iov.iov_base = (void *) buf;
-	msg_iov.iov_len = len;
-	rma_iov.addr = addr;
-	rma_iov.len = len;
-	rma_iov.key = key;
-
-	return smr_generic_rma(ep_fid, &msg_iov, 1, &rma_iov, 1, &desc, 
-			       dest_addr, context, ofi_op_write, 0, 0);
-}
-
-ssize_t smr_writev(struct fid_ep *ep_fid, const struct iovec *iov, void **desc,
-		size_t count, fi_addr_t dest_addr, uint64_t addr, uint64_t key,
-		void *context)
-{
-	struct fi_rma_iov rma_iov;
-
-	rma_iov.addr = addr;
-	rma_iov.len  = ofi_total_iov_len(iov, count);
-	rma_iov.key = key;
-
-	return smr_generic_rma(ep_fid, iov, count, &rma_iov, 1, desc,
-			       dest_addr, context, ofi_op_write, 0, 0);
-}
-
-
-ssize_t smr_writemsg(struct fid_ep *ep_fid, const struct fi_msg_rma *msg,
-	uint64_t flags)
-{
-	return smr_generic_rma(ep_fid, msg->msg_iov, msg->iov_count,
-			       msg->rma_iov, msg->rma_iov_count,
-			       msg->desc, msg->addr, msg->context,
-			       ofi_op_write, msg->data,
-			       (flags & FI_REMOTE_CQ_DATA) ?
-			       OFI_REMOTE_CQ_DATA : 0);
-}
-
-ssize_t smr_rma_inject(struct fid_ep *ep_fid, const void *buf, size_t len,
-	fi_addr_t dest_addr, uint64_t addr, uint64_t key)
-{
-	struct smr_ep *ep;
-	struct smr_region *peer_smr;
-	struct smr_cmd *cmd;
-	struct iovec iov, rma_iovec;
-	struct fi_rma_iov rma_iov;
-	int peer_id;
-	ssize_t ret = 0;
-
-	assert(len <= SMR_INJECT_SIZE);
-
-	ep = container_of(ep_fid, struct smr_ep, util_ep.ep_fid.fid);
-
-	peer_id = (int) dest_addr;
-	ret = smr_verify_peer(ep, peer_id);
-	if (ret)
-		return ret;
-
-	peer_smr = smr_peer_region(ep->region, peer_id);
-	fastlock_acquire(&peer_smr->lock);
-	if (!peer_smr->cmd_cnt) {
-		ret = -FI_EAGAIN;
-		goto unlock_region;
-	}
-
-	iov.iov_base = (void *) buf;
-	iov.iov_len = len;
-	rma_iovec.iov_base = (void *) addr;
-	rma_iovec.iov_len = len;
-
-	ret = process_vm_writev(peer_smr->pid, &iov, 1,
-				&rma_iovec, 1, 0);
-
-	if (ret != len) {
-		if (ret < 0) {
-			FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-				"CMA write error\n");
-			ret = -errno;
-		} else {
-			FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-				"unable to process tx completion\n");
-			ret = -FI_EIO;
-		}
-		goto unlock_region;
-	}
-	ret = 0;
-
-	rma_iov.addr = addr;
-	rma_iov.len = len;
-
-	cmd = ofi_cirque_tail(smr_cmd_queue(peer_smr));
-	smr_format_rma_resp(cmd, peer_id, &rma_iov, 1, len,
-			    ofi_op_read_rsp);
-	ofi_cirque_commit(smr_cmd_queue(peer_smr));
-	peer_smr->cmd_cnt--;
-
-unlock_region:
-	fastlock_release(&peer_smr->lock);
-	return ret;
-}
-
-ssize_t smr_writedata(struct fid_ep *ep, const void *buf, size_t len,
-		      void *desc, uint64_t data, fi_addr_t dest_addr,
-		      uint64_t addr, uint64_t key, void *context)
-{
-	struct iovec iov;
-	struct fi_rma_iov rma_iov;
-
-	iov.iov_base = (void *) buf;
-	iov.iov_len = len;
-	rma_iov.addr = addr;
-	rma_iov.len = len;
-	rma_iov.key = key;
-
-	return smr_generic_rma(ep, &iov, 1, &rma_iov, 1, &desc, dest_addr, context,
-			       ofi_op_write, data, OFI_REMOTE_CQ_DATA);
-}
-
-ssize_t smr_inject_writedata(struct fid_ep *ep_fid, const void *buf, size_t len,
-			     uint64_t data, fi_addr_t dest_addr, uint64_t addr,
-			     uint64_t key)
-{
-	struct smr_ep *ep;
-	struct smr_region *peer_smr;
-	struct smr_inject_buf *tx_buf;
-	struct smr_cmd *cmd;
-	struct iovec iov;
-	struct fi_rma_iov rma_iov;
-	int peer_id;
-	ssize_t ret = 0;
-
-	assert(len <= SMR_INJECT_SIZE);
-	ep = container_of(ep_fid, struct smr_ep, util_ep.ep_fid.fid);
-
-	peer_id = (int) dest_addr;
-	ret = smr_verify_peer(ep, peer_id);
-	if (ret)
-		return ret;
-
-	peer_smr = smr_peer_region(ep->region, peer_id);
-	fastlock_acquire(&peer_smr->lock);
-	if (peer_smr->cmd_cnt < 2) {
-		ret = -FI_EAGAIN;
-		goto unlock_region;
-	}
-
-	iov.iov_base = (void *) buf;
-	iov.iov_len = len;
-	rma_iov.addr = addr;
-	rma_iov.len = len;
-	rma_iov.key = key;
-
-	cmd = ofi_cirque_tail(smr_cmd_queue(peer_smr));
-
-	if (len <= SMR_MSG_DATA_LEN) {
-		smr_format_inline(cmd, smr_peer_addr(ep->region)[peer_id].addr,
-				  &iov, 1, NULL, ofi_op_write, 0, data,
-				  OFI_REMOTE_CQ_DATA);
-	} else {
-		tx_buf = smr_freestack_pop(smr_inject_pool(peer_smr));
-		smr_format_inject(cmd, smr_peer_addr(ep->region)[peer_id].addr, &iov,
-				  1, NULL, ofi_op_write, 0, data, OFI_REMOTE_CQ_DATA,
-				  peer_smr, tx_buf);
-	}
-
-	ofi_cirque_commit(smr_cmd_queue(peer_smr));
-	peer_smr->cmd_cnt--;
-	cmd = ofi_cirque_tail(smr_cmd_queue(peer_smr));
-	smr_format_rma_iov(cmd, &rma_iov, 1);
-	ofi_cirque_commit(smr_cmd_queue(peer_smr));
-	peer_smr->cmd_cnt--;
-
-unlock_region:
-	fastlock_release(&peer_smr->lock);
-	return ret;
-}
-
-static struct fi_ops_rma smr_rma_ops = {
-	.size = sizeof(struct fi_ops_rma),
-	.read = smr_read,
-	.readv = smr_readv,
-	.readmsg = smr_readmsg,
-	.write = smr_write,
-	.writev = smr_writev,
-	.writemsg = smr_writemsg,
-	.inject = smr_rma_inject,
-	.writedata = smr_writedata,
-	.injectdata = smr_inject_writedata,
-};
-
-static void smr_ioc_to_iov(const struct fi_ioc *ioc, struct iovec *iov,
-			   size_t count, size_t size)
-{
-	int i;
-	for (i = 0; i < count; i++) {
-		iov[i].iov_base = ioc[i].addr;
-		iov[i].iov_len = ioc[i].count * size;
-	}
-}
-
-static int smr_fetch_result(struct smr_ep *ep, struct smr_region *peer_smr,
-			    struct iovec *iov, size_t iov_count,
-			    const struct fi_rma_ioc *rma_ioc, size_t rma_count,
-			    enum fi_datatype datatype, size_t total_len)
-{
-	int ret, i;
-	struct iovec rma_iov[SMR_IOV_LIMIT];
-
-	for (i = 0; i < rma_count; i++) {
-		rma_iov[i].iov_base = (void *) rma_ioc[i].addr;
-		rma_iov[i].iov_len = rma_ioc[i].count * ofi_datatype_size(datatype);
-	}
-
-	ret = process_vm_readv(peer_smr->pid, iov, iov_count,
-			       rma_iov, rma_count, 0);
-	if (ret != total_len) {
-		if (ret < 0) {
-			FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-				"CMA write error\n");
-			return -errno;
-		} else { 
-			FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-				"partial read occurred\n");
-			return -FI_EIO;
-		}
-	}
-
-	return 0; 
-}
-
-static ssize_t smr_generic_atomic(struct fid_ep *ep_fid,
-			const struct fi_ioc *ioc, void **desc, size_t count,
-			const struct fi_ioc *compare_ioc, void **compare_desc,
-			size_t compare_count, struct fi_ioc *result_ioc,
-			void **result_desc, size_t result_count,
-			fi_addr_t addr, const struct fi_rma_ioc *rma_ioc,
-			size_t rma_count, enum fi_datatype datatype,
-			enum fi_op atomic_op, void *context, uint32_t op)
-{
-	struct smr_ep *ep;
-	struct smr_region *peer_smr;
-	struct smr_inject_buf *tx_buf;
-	struct smr_cmd *cmd;
-	struct iovec iov[SMR_IOV_LIMIT];
-	struct iovec compare_iov[SMR_IOV_LIMIT];
-	struct iovec result_iov[SMR_IOV_LIMIT];
-	int peer_id, err = 0;
-	ssize_t ret = 0;
-	size_t msg_len, total_len;
-
-	assert(count <= SMR_IOV_LIMIT);
-	assert(result_count <= SMR_IOV_LIMIT);
-	assert(compare_count <= SMR_IOV_LIMIT);
-	assert(rma_count <= SMR_IOV_LIMIT);
-
-	ep = container_of(ep_fid, struct smr_ep, util_ep.ep_fid.fid);
-
-	peer_id = (int) addr;
-	ret = smr_verify_peer(ep, peer_id);
-	if(ret)
-		return ret;
-
-	peer_smr = smr_peer_region(ep->region, peer_id);
-	fastlock_acquire(&peer_smr->lock);
-	if (peer_smr->cmd_cnt < 2) {
-		ret = -FI_EAGAIN;
-		goto unlock_region;
-	}
-
-	fastlock_acquire(&ep->util_ep.tx_cq->cq_lock);
-	if (ofi_cirque_isfull(ep->util_ep.tx_cq->cirq)) {
-		ret = -FI_EAGAIN;
-		goto unlock_cq;
-	}
-
-	cmd = ofi_cirque_tail(smr_cmd_queue(peer_smr));
-	msg_len = total_len = ofi_datatype_size(datatype) *
-			      ofi_total_ioc_cnt(ioc, count);
-	
-	switch (op) {
-	case ofi_op_atomic_compare:
-		if (!compare_ioc) {
-			FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-				"no compare buffer given\n");
-			ret = -FI_EINVAL;
-			goto unlock_cq;
-		}
-		smr_ioc_to_iov(compare_ioc, compare_iov, compare_count,
-			       ofi_datatype_size(datatype));
-		total_len *= 2;
-		/* fall through */
-	case ofi_op_atomic_fetch:
-		if (!result_ioc) {
-			FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-				"no result buffer given\n");
-			ret = -FI_EINVAL;
-			goto unlock_cq;
-		}
-		smr_ioc_to_iov(result_ioc, result_iov, result_count,
-			       ofi_datatype_size(datatype));
-		/* fall through */
-	case ofi_op_atomic:
-		if (!ioc && atomic_op != FI_ATOMIC_READ) {
-			FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-				"no atomic buffer given\n");
-			ret = -FI_EINVAL;
-			goto unlock_cq;
-		}
-		smr_ioc_to_iov(ioc, iov, count, ofi_datatype_size(datatype));
-		break;
-	default:
-		break;
-	}
-
-	if (total_len <= SMR_MSG_DATA_LEN) {
-		smr_format_inline_atomic(cmd, smr_peer_addr(ep->region)[peer_id].addr,
-					 iov, count, compare_iov, compare_count,
-					 context, op, datatype, atomic_op);
-	} else if (total_len <= SMR_INJECT_SIZE) {
-		tx_buf = smr_freestack_pop(smr_inject_pool(peer_smr));
-		smr_format_inject_atomic(cmd, smr_peer_addr(ep->region)[peer_id].addr,
-					 iov, count, compare_iov, compare_count,
-					 context, op, datatype, atomic_op,
-					 peer_smr, tx_buf);
-	} else {
-		FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-			"message too large\n");
-		ret = -FI_EINVAL;
-		goto unlock_cq;
-	}
-
-	ofi_cirque_commit(smr_cmd_queue(peer_smr));
-	peer_smr->cmd_cnt--;
-	cmd = ofi_cirque_tail(smr_cmd_queue(peer_smr));
-	smr_format_rma_ioc(cmd, rma_ioc, rma_count);
-	ofi_cirque_commit(smr_cmd_queue(peer_smr));
-	peer_smr->cmd_cnt--;
-
-	if (op != ofi_op_atomic) {
-		err = smr_fetch_result(ep, peer_smr, result_iov, result_count,
-				       rma_ioc, rma_count, datatype, msg_len);
-		if (err)
-			FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-				"unable to fetch results");
-	}
-
-	ret = ep->tx_comp(ep, context, smr_tx_comp_flags(op), err);
-	if (ret) {
-		FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-			"unable to process tx completion\n");
-	}
-
-unlock_cq:
-	fastlock_release(&ep->util_ep.tx_cq->cq_lock);
-unlock_region:
-	fastlock_release(&peer_smr->lock);
-	return ret;
-}
-
-static ssize_t smr_atomic_writemsg(struct fid_ep *ep,
-			const struct fi_msg_atomic *msg, uint64_t flags)
-{
-	return smr_generic_atomic(ep, msg->msg_iov, msg->desc, msg->iov_count,
-				  NULL, NULL, 0, NULL, NULL, 0, msg->addr,
-				  msg->rma_iov, msg->rma_iov_count,
-				  msg->datatype, msg->op, msg->context,
-				  ofi_op_atomic);
-}
-
-static ssize_t smr_atomic_writev(struct fid_ep *ep,
-			const struct fi_ioc *iov, void **desc, size_t count,
-			fi_addr_t dest_addr, uint64_t addr, uint64_t key,
-			enum fi_datatype datatype, enum fi_op op, void *context)
-{
-	struct fi_rma_ioc rma_iov;
-
-	rma_iov.addr = addr;
-	rma_iov.count = ofi_total_ioc_cnt(iov, count);
-	rma_iov.key = key;
-
-	return smr_generic_atomic(ep, iov, desc, count, NULL, NULL, 0, NULL,
-				  NULL, 0, dest_addr, &rma_iov, 1, datatype,
-				  op, context, ofi_op_atomic);
-}
-
-static ssize_t smr_atomic_write(struct fid_ep *ep, const void *buf, size_t count,
-			void *desc, fi_addr_t dest_addr, uint64_t addr,
-			uint64_t key, enum fi_datatype datatype, enum fi_op op,
-			void *context)
-{
-	struct fi_ioc iov;
-	struct fi_rma_ioc rma_iov;
-
-	iov.addr = (void *) buf;
-	iov.count = count;
-
-	rma_iov.addr = addr;
-	rma_iov.count = count;
-	rma_iov.key = key;
-
-	return smr_generic_atomic(ep, &iov, &desc, 1, NULL, NULL, 0, NULL, NULL, 0,
-				  dest_addr, &rma_iov, 1, datatype, op, context,
-				  ofi_op_atomic);
-}
-
-static ssize_t smr_atomic_inject(struct fid_ep *ep_fid, const void *buf,
-			size_t count, fi_addr_t dest_addr, uint64_t addr,
-			uint64_t key, enum fi_datatype datatype, enum fi_op op)
-{
-	struct smr_ep *ep;
-	struct smr_region *peer_smr;
-	struct smr_inject_buf *tx_buf;
-	struct smr_cmd *cmd;
-	struct iovec iov;
-	struct fi_rma_ioc rma_ioc;
-	int peer_id;
-	ssize_t ret = 0;
-	size_t total_len;
-
-	assert(count <= SMR_INJECT_SIZE);
-
-	ep = container_of(ep_fid, struct smr_ep, util_ep.ep_fid.fid);
-
-	peer_id = (int) dest_addr;
-	ret = smr_verify_peer(ep, peer_id);
-	if(ret)
-		return ret;
-
-	peer_smr = smr_peer_region(ep->region, peer_id);
-	fastlock_acquire(&peer_smr->lock);
-	if (peer_smr->cmd_cnt < 2) {
-		ret = -FI_EAGAIN;
-		goto unlock_region;
-	}
-
-	cmd = ofi_cirque_tail(smr_cmd_queue(peer_smr));
-	total_len = count * ofi_datatype_size(datatype);
-	
-	iov.iov_base = (void *) buf;
-	iov.iov_len = total_len;
-
-	rma_ioc.addr = addr;
-	rma_ioc.count = count;
-	rma_ioc.key = key;
-
-	if (total_len <= SMR_MSG_DATA_LEN) {
-		smr_format_inline_atomic(cmd, smr_peer_addr(ep->region)[peer_id].addr,
-					 &iov, 1, NULL, 0, NULL, ofi_op_atomic,
-					 datatype, op);
-	} else if (total_len <= SMR_INJECT_SIZE) {
-		tx_buf = smr_freestack_pop(smr_inject_pool(peer_smr));
-		smr_format_inject_atomic(cmd, smr_peer_addr(ep->region)[peer_id].addr,
-					 &iov, 1, NULL, 0, NULL, ofi_op_atomic,
-					 datatype, op, peer_smr, tx_buf);
-	}
-
-	ofi_cirque_commit(smr_cmd_queue(peer_smr));
-	peer_smr->cmd_cnt--;
-	cmd = ofi_cirque_tail(smr_cmd_queue(peer_smr));
-	smr_format_rma_ioc(cmd, &rma_ioc, 1);
-	ofi_cirque_commit(smr_cmd_queue(peer_smr));
-	peer_smr->cmd_cnt--;
-
-unlock_region:
-	fastlock_release(&peer_smr->lock);
-	return ret;
-}
-
-static ssize_t smr_atomic_readwritemsg(struct fid_ep *ep,
-			const struct fi_msg_atomic *msg, struct fi_ioc *resultv,
-			void **result_desc, size_t result_count, uint64_t flags)
-{
-	return smr_generic_atomic(ep, msg->msg_iov, msg->desc, msg->iov_count,
-				  NULL, NULL, 0, resultv, result_desc,
-				  result_count, msg->addr,
-				  msg->rma_iov, msg->rma_iov_count,
-				  msg->datatype, msg->op, msg->context,
-				  ofi_op_atomic_fetch);
-}
-
-static ssize_t smr_atomic_readwritev(struct fid_ep *ep,
-			const struct fi_ioc *iov, void **desc, size_t count,
-			struct fi_ioc *resultv, void **result_desc,
-			size_t result_count, fi_addr_t dest_addr, uint64_t addr,
-			uint64_t key, enum fi_datatype datatype, enum fi_op op,
-			void *context)
-{
-	struct fi_rma_ioc rma_iov;
-
-	rma_iov.addr = addr;
-	rma_iov.count = ofi_total_ioc_cnt(iov, count);
-	rma_iov.key = key;
-
-	return smr_generic_atomic(ep, iov, desc, count, NULL, NULL, 0, resultv,
-				  result_desc, result_count, dest_addr,
-				  &rma_iov, 1, datatype, op, context,
-				  ofi_op_atomic_fetch);
-}
-
-static ssize_t smr_atomic_readwrite(struct fid_ep *ep, const void *buf,
-			size_t count, void *desc, void *result,
-			void *result_desc, fi_addr_t dest_addr, uint64_t addr,
-			uint64_t key, enum fi_datatype datatype, enum fi_op op,
-			void *context)
-{
-	struct fi_ioc iov, resultv;
-	struct fi_rma_ioc rma_iov;
-
-	iov.addr = (void *) buf;
-	iov.count = count;
-
-	resultv.addr = result;
-	resultv.count = count;
-
-	rma_iov.addr = addr;
-	rma_iov.count = count;
-	rma_iov.key = key;
-
-	return smr_generic_atomic(ep, &iov, &desc, 1, NULL, NULL, 0, &resultv,
-				  &result_desc, 1, dest_addr, &rma_iov, 1,
-				  datatype, op, context, ofi_op_atomic_fetch);
-}
-
-static ssize_t smr_atomic_compwritemsg(struct fid_ep *ep,
-			const struct fi_msg_atomic *msg,
-			const struct fi_ioc *comparev, void **compare_desc,
-			size_t compare_count, struct fi_ioc *resultv,
-			void **result_desc, size_t result_count, uint64_t flags)
-{
-	return smr_generic_atomic(ep, msg->msg_iov, msg->desc, msg->iov_count,
-				  comparev, compare_desc, compare_count,
-				  resultv, result_desc,
-				  result_count, msg->addr,
-				  msg->rma_iov, msg->rma_iov_count,
-				  msg->datatype, msg->op, msg->context,
-				  ofi_op_atomic_compare);
-}
-
-static ssize_t smr_atomic_compwritev(struct fid_ep *ep,
-			const struct fi_ioc *iov, void **desc, size_t count,
-			const struct fi_ioc *comparev, void **compare_desc,
-			size_t compare_count, struct fi_ioc *resultv,
-			void **result_desc, size_t result_count,
-			fi_addr_t dest_addr, uint64_t addr, uint64_t key,
-			enum fi_datatype datatype, enum fi_op op, void *context)
-{
-	struct fi_rma_ioc rma_iov;
-
-	rma_iov.addr = addr;
-	rma_iov.count = ofi_total_ioc_cnt(iov, count);
-	rma_iov.key = key;
-
-	return smr_generic_atomic(ep, iov, desc, count, comparev, compare_desc,
-				  compare_count, resultv, result_desc,
-				  result_count, dest_addr, &rma_iov, 1,
-				  datatype, op, context, ofi_op_atomic_compare);
-}
-
-static ssize_t smr_atomic_compwrite(struct fid_ep *ep, const void *buf,
-			size_t count, void *desc, const void *compare,
-			void *compare_desc, void *result, void *result_desc,
-			fi_addr_t dest_addr, uint64_t addr, uint64_t key,
-			enum fi_datatype datatype, enum fi_op op, void *context)
-{
-	struct fi_ioc iov, resultv, comparev;
-	struct fi_rma_ioc rma_iov;
-
-	iov.addr = (void *) buf;
-	iov.count = count;
-
-	resultv.addr = result;
-	resultv.count = count;
-
-	comparev.addr = (void *) compare;
-	comparev.count = count;
-
-	rma_iov.addr = addr;
-	rma_iov.count = count;
-	rma_iov.key = key;
-
-	return smr_generic_atomic(ep, &iov, &desc, 1, &comparev, &compare_desc,
-				  1, &resultv, &result_desc, 1, dest_addr,
-				  &rma_iov, 1, datatype, op, context,
-				  ofi_op_atomic_compare);
-}
-
-int smr_query_atomic(struct fid_domain *domain, enum fi_datatype datatype,
-		     enum fi_op op, struct fi_atomic_attr *attr, uint64_t flags)
-{
-	int ret;
-	size_t total_size;
-
-	if (flags & FI_TAGGED) {
-		FI_WARN(&smr_prov, FI_LOG_EP_CTRL,
-			"tagged atomic op not supported\n");
-		return -FI_EINVAL;
-	}
-
-	ret = ofi_atomic_valid(&smr_prov, datatype, op, flags);
-	if (ret || !attr)
-		return ret;
-
-	attr->size = ofi_datatype_size(datatype);
-
-	total_size = (flags & FI_COMPARE_ATOMIC) ? SMR_COMP_INJECT_SIZE :
-		      SMR_INJECT_SIZE;
-	attr->count = total_size / attr->size;
-
-	return ret;
-}
-
-static int smr_atomic_valid(struct fid_ep *ep, enum fi_datatype datatype,
-			    enum fi_op op, size_t *count)
-{
-	struct fi_atomic_attr attr;
-	int ret;
-
-	ret = smr_query_atomic(NULL, datatype, op, &attr, 0);
-
-	if (!ret)
-		*count = attr.count;
-
-	return ret;
-}
-
-static int smr_atomic_fetch_valid(struct fid_ep *ep, enum fi_datatype datatype,
-				  enum fi_op op, size_t *count)
-{
-	struct fi_atomic_attr attr;
-	int ret;
-
-	ret = smr_query_atomic(NULL, datatype, op, &attr, FI_FETCH_ATOMIC);
-
-	if (!ret)
-		*count = attr.count;
-
-	return ret;
-}
-
-static int smr_atomic_comp_valid(struct fid_ep *ep, enum fi_datatype datatype,
-				 enum fi_op op, size_t *count)
-{
-	struct fi_atomic_attr attr;
-	int ret;
-
-	ret = smr_query_atomic(NULL, datatype, op, &attr, FI_COMPARE_ATOMIC);
-
-	if (!ret)
-		*count = attr.count;
-
-	return ret;
-}
-
-struct fi_ops_atomic smr_atomic_ops = {
-	.size = sizeof(struct fi_ops_atomic),
-	.write = smr_atomic_write,
-	.writev = smr_atomic_writev,
-	.writemsg = smr_atomic_writemsg,
-	.inject = smr_atomic_inject,
-	.readwrite = smr_atomic_readwrite,
-	.readwritev = smr_atomic_readwritev,
-	.readwritemsg = smr_atomic_readwritemsg,
-	.compwrite = smr_atomic_compwrite,
-	.compwritev = smr_atomic_compwritev,
-	.compwritemsg = smr_atomic_compwritemsg,
-	.writevalid = smr_atomic_valid,
-	.readwritevalid = smr_atomic_fetch_valid,
-	.compwritevalid = smr_atomic_comp_valid,
-};
 
 static int smr_ep_close(struct fid *fid)
 {
